fix registerRestRoute types
